Moving project objects and tests to Project Service/DAO
