+++ digiroad2-viite/src/test/scala/fi/liikennevirasto/viite/dao/RoadAddressDAOSpec.scala
@@ -165,42 +165,4 @@ class RoadAddressDAOSpec extends FunSuite with Matchers {
     }
   }
 
-  test("create empty road address project") {
-    runWithRollback {
-      val id = Sequences.nextViitePrimaryKeySeqValue
-      val rap = RoadAddressProject(id, ProjectState.apply(1), "TestProject", "TestUser", DateTime.parse("1901-01-01"), "TestUser", DateTime.parse("1901-01-01"), DateTime.now(), "Some additional info", List.empty[ReservedRoadPart])
-      ProjectDAO.createRoadAddressProject(rap)
-      ProjectDAO.getRoadAddressProjectById(id).nonEmpty should be(true)
-    }
-  }
-
-
-  test("create road address project with project links") {
-    def toProjectLink(project: RoadAddressProject)(roadAddress: RoadAddress): ProjectLink = {
-      ProjectLink(id=NewRoadAddress, roadAddress.roadNumber, roadAddress.roadPartNumber, roadAddress.track,
-        roadAddress.discontinuity, roadAddress.startAddrMValue, roadAddress.endAddrMValue, roadAddress.startDate,
-        roadAddress.endDate, modifiedBy=Option(project.createdBy), 0L, roadAddress.linkId, roadAddress.startMValue, roadAddress.endMValue,
-        roadAddress.sideCode, roadAddress.calibrationPoints, floating=false, roadAddress.geom, project.id, LinkStatus.NotHandled)
-    }
-    val address=ReservedRoadPart(5:Long, 203:Long, 203:Long, 5.5:Double, Discontinuity.apply("jatkuva"), 8:Long)
-    runWithRollback {
-      val id = Sequences.nextViitePrimaryKeySeqValue
-      val rap = RoadAddressProject(id, ProjectState.apply(1), "TestProject", "TestUser", DateTime.parse("1901-01-01"), "TestUser", DateTime.parse("1901-01-01"), DateTime.now(), "Some additional info", List(address))
-      ProjectDAO.createRoadAddressProject(rap)
-      val addresses = RoadAddressDAO.fetchByRoadPart(5, 203).map(toProjectLink(rap))
-      ProjectDAO.create(addresses)
-      ProjectDAO.getRoadAddressProjectById(id).nonEmpty should be(true)
-      val projectlinks=ProjectDAO.getProjectLinks(id)
-      projectlinks.length should be > 0
-    }
-  }
-
-  test("get roadpart info") {
-    runWithRollback {
-      val reserveResult= RoadAddressDAO.getRoadPartInfo(5,203)
-      val expectedLink = reserveResult==Some((242,5172706,5907.0,5))
-      expectedLink should be (true)
-    }
-
-  }
 }
